+++ tests/AltaSoft.Simpra.tests/Models/TestModelFactory.cs
@@ -1,0 +1,9 @@
+namespace AltaSoft.Simpra.Tests.Models;
+
+internal static class TestModelFactory
+{
+    public static TestModel GetTestModel()
+    {
+        return new TestModel { Transfer = new Transfer { Amount = 100, Currency = "USD" }, Customer = new Customer { Id = 1, Status = 1 }, Remittance = "Test" };
+    }
+}
